feat(0.4.0): refactor spec transaction types
Spec version 0.3.0 distinguished between two types of transactions:
those in a block and those sent out by clients. As of 0.4.0, there was
an attempt to converge into a single transaction type. However, this
is problematic as they're inherently different. This commit undoes that
change and brings back `BROADCASTED_*` types.

This commit was made by simply refactoring the spec backward, as well as
fixing some bugs (?) introduced in the original unification process:

    1. Query-only tx versions incorrectly disallowed.
    2. Simulating/estimating deprecated tx types incorrectly allowed.

On top of this, the `oneOf`s in `DECLARE_TXN_V*` have been flattened
based on the `description` content, enhancing type safety.

Otherwise, as with all other patches, this commit makes no changes to
any data structure sent over the network.