feat: support deser-only fields
Some fields have a fixed value and must be present when deserializing
just to verify that the value is indeed the expected fixed one (it's
never part of the struct). This commit adds support for it via a new
`must_present_in_deser` profile option.

Strictly speaking, fields like `type` and `version` also must exist.
However, our hands are tied there due to the current deserliazation
implementation on enum-wrapped struct types - when deserializing through
the enum type, the tag fields are consumed and no longer available, so
the lower struct types must be able to handle their absence.