feat(0.4.0): patch trace spec for optional fields
In this patch, whether a field is optional is derived from actual node
implementations (pathfinder & juno). It's hard to say whether this patch
is correct or the nodes have a bug for not always sending these fields,
but let's settle with this for now as v0.5.0 will correctly mark the
fields anyways.