fix(0.8.0): refactor spec to remove `transaction_hash` on block receipt txs
Previously, this repo made a patch to the spec merging the
`transaction_hash` field into the `TXN` type. This patch resulted in
easier to use generated types, and it worked because at that time, the
type `TXN` is _always_ used together with the `transaction_hash` field
(via `allOf`), allowing the wire format to remain compliant despite
the patch.

However, since v0.7.x, the upstream spec started using the `TXN`
type, specifically in the `BLOCK_BODY_WITH_RECEIPTS` type. This change
broke the assumption of the patch. Nevertheless, this issue was not
caught during the v0.7.x time as `pathfinder` still served the field,
essentially deviating from the spec:

    https://github.com/eqlabs/pathfinder/blob/6d5941700e783af7fb7633c3155014f660a296ac/crates/rpc/src/method/get_block_with_receipts.rs#L160-L165

`pathfinder` no longer does that for v0.8.x, and hence the issue is now
exposed. This commit partially undoes the patch to be compliant again.